feat: refactor socket.io integration and move handling to separate module
